Add Spark VPJ quota tests
Cover refreshed pre-write storage quota enforcement, KIF skip behavior, and
VPJ quota status handling.

Split VenicePushJobTest into thematic test classes with shared fixtures.
