+++ clients/venice-push-job/src/test/java/com/linkedin/venice/spark/datawriter/jobs/AbstractDataWriterSparkJobTest.java
@@ -6,6 +6,8 @@ import static com.linkedin.venice.spark.SparkConstants.CHUNKED_KEY_SUFFIX_COLUMN
 import static com.linkedin.venice.spark.SparkConstants.KEY_COLUMN_NAME;
 import static com.linkedin.venice.spark.SparkConstants.MESSAGE_TYPE_COLUMN_NAME;
 import static com.linkedin.venice.spark.SparkConstants.OFFSET_COLUMN_NAME;
+import static com.linkedin.venice.spark.SparkConstants.PARTITION_COLUMN_NAME;
+import static com.linkedin.venice.spark.SparkConstants.RAW_PUBSUB_INPUT_TABLE_SCHEMA;
 import static com.linkedin.venice.spark.SparkConstants.RMD_COLUMN_NAME;
 import static com.linkedin.venice.spark.SparkConstants.RMD_VERSION_ID_COLUMN_NAME;
 import static com.linkedin.venice.spark.SparkConstants.SCHEMA_ID_COLUMN_NAME;
@@ -15,6 +17,7 @@ import static com.linkedin.venice.spark.SparkConstants.SPARK_SESSION_CONF_PREFIX
 import static com.linkedin.venice.spark.SparkConstants.VALUE_COLUMN_NAME;
 import static com.linkedin.venice.vpj.VenicePushJobConstants.DEFAULT_KEY_FIELD_PROP;
 import static com.linkedin.venice.vpj.VenicePushJobConstants.DEFAULT_VALUE_FIELD_PROP;
+import static com.linkedin.venice.vpj.VenicePushJobConstants.SPARK_NATIVE_INPUT_FORMAT_ENABLED;
 import static org.apache.spark.sql.types.DataTypes.BinaryType;
 import static org.apache.spark.sql.types.DataTypes.IntegerType;
 import static org.apache.spark.sql.types.DataTypes.LongType;
@@ -29,9 +32,12 @@ import com.linkedin.venice.etl.ETLValueSchemaTransformation;
 import com.linkedin.venice.exceptions.VeniceException;
 import com.linkedin.venice.hadoop.PushJobSetting;
 import com.linkedin.venice.hadoop.exceptions.VeniceInvalidInputException;
+import com.linkedin.venice.hadoop.exceptions.VeniceStorageQuotaExceededException;
 import com.linkedin.venice.jobs.ComputeJob;
 import com.linkedin.venice.jobs.DataWriterComputeJob;
+import com.linkedin.venice.kafka.protocol.enums.MessageType;
 import com.linkedin.venice.meta.Version;
+import com.linkedin.venice.meta.VersionImpl;
 import com.linkedin.venice.partitioner.DefaultVenicePartitioner;
 import com.linkedin.venice.schema.rmd.RmdSchemaGenerator;
 import com.linkedin.venice.serializer.FastSerializerDeserializerFactory;
@@ -44,23 +50,29 @@ import com.linkedin.venice.utils.VeniceProperties;
 import java.io.File;
 import java.io.IOException;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 import java.util.Properties;
+import java.util.concurrent.atomic.AtomicLong;
 import org.apache.avro.Schema;
 import org.apache.avro.generic.GenericData;
 import org.apache.avro.generic.GenericRecord;
 import org.apache.hadoop.fs.Path;
 import org.apache.spark.api.java.JavaSparkContext;
+import org.apache.spark.api.java.function.MapPartitionsFunction;
+import org.apache.spark.broadcast.Broadcast;
 import org.apache.spark.rdd.RDD;
 import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Row;
 import org.apache.spark.sql.RowFactory;
 import org.apache.spark.sql.RuntimeConfig;
 import org.apache.spark.sql.SparkSession;
+import org.apache.spark.sql.catalyst.expressions.GenericRowWithSchema;
 import org.apache.spark.sql.types.Metadata;
 import org.apache.spark.sql.types.StructField;
 import org.apache.spark.sql.types.StructType;
 import org.testng.Assert;
+import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 
@@ -352,6 +364,118 @@ public class AbstractDataWriterSparkJobTest {
     }
   }
 
+  @DataProvider(name = "storageQuotaMatrix")
+  public Object[][] storageQuotaMatrix() {
+    return new Object[][] {
+        { "big file, small start quota, changed to big before write", 1L, StorageQuotaRefresh.BIG, true, true },
+        { "big file, small start quota, no change", 1L, StorageQuotaRefresh.SMALL, true, false },
+        { "small file, big start quota, no change", Long.MAX_VALUE, StorageQuotaRefresh.BIG, false, true },
+        { "small file, big start quota, changed to small before write", Long.MAX_VALUE, StorageQuotaRefresh.SMALL,
+            false, false },
+        { "big file, small start quota, changed to exactly input size before write", 1L,
+            StorageQuotaRefresh.EXACT_INPUT_SIZE, true, true } };
+  }
+
+  @Test(dataProvider = "storageQuotaMatrix")
+  public void testRunJobEnforcesPreWriteStorageQuotaMatrix(
+      String caseName,
+      long startQuota,
+      StorageQuotaRefresh refreshedQuota,
+      boolean inputShouldExceedStartQuota,
+      boolean expectSuccess) throws IOException {
+    File inputDir = TestWriteUtils.getTempDataDirectory();
+    Schema dataSchema = TestWriteUtils.writeSimpleAvroFileWithStringToStringSchema(inputDir);
+    PushJobSetting setting = getDefaultPushJobSetting(inputDir, dataSchema);
+    setting.partitionCount = 1;
+    setting.rmdField = "";
+    setting.storeStorageQuota = startQuota;
+
+    Properties properties = new Properties();
+    properties.setProperty(SPARK_NATIVE_INPUT_FORMAT_ENABLED, String.valueOf(true));
+    AtomicLong quotaRefreshCount = new AtomicLong();
+
+    try (QuotaTestingDataWriterSparkJob job = new QuotaTestingDataWriterSparkJob()) {
+      job.configure(new VeniceProperties(properties), setting);
+      job.setCurrentStorageQuotaSupplier(() -> {
+        quotaRefreshCount.incrementAndGet();
+        return refreshedQuota.resolve(job);
+      });
+
+      job.runJob();
+
+      Assert.assertEquals(quotaRefreshCount.get(), 1L);
+      long totalInputDataSize = getTotalInputDataSize(job);
+      Assert.assertTrue(totalInputDataSize > 1L, caseName);
+      if (inputShouldExceedStartQuota) {
+        Assert.assertTrue(totalInputDataSize > startQuota, caseName);
+      } else {
+        Assert.assertTrue(totalInputDataSize < startQuota, caseName);
+      }
+
+      if (expectSuccess) {
+        Assert.assertEquals(job.getStatus(), ComputeJob.Status.SUCCEEDED, caseName);
+        Assert.assertNull(job.getFailureReason(), caseName);
+        Assert.assertTrue(job.isWriterFactoryCreated(), caseName);
+      } else {
+        Assert.assertEquals(job.getStatus(), ComputeJob.Status.FAILED, caseName);
+        Assert.assertTrue(job.getFailureReason() instanceof VeniceStorageQuotaExceededException, caseName);
+        Assert.assertFalse(job.isWriterFactoryCreated(), caseName);
+      }
+    }
+  }
+
+  @Test
+  public void testRunJobSkipsPreWriteStorageQuotaCheckForKifRepush() throws IOException {
+    PushJobSetting setting = getDefaultKafkaInputPushJobSetting();
+    setting.storeStorageQuota = 1L;
+
+    AtomicLong quotaRefreshCount = new AtomicLong();
+    try (KifQuotaTestingDataWriterSparkJob job = new KifQuotaTestingDataWriterSparkJob()) {
+      job.configure(new VeniceProperties(new Properties()), setting);
+      job.setCurrentStorageQuotaSupplier(() -> {
+        quotaRefreshCount.incrementAndGet();
+        return 1L;
+      });
+
+      job.runJob();
+
+      Assert.assertEquals(job.getStatus(), ComputeJob.Status.SUCCEEDED);
+      Assert.assertNull(job.getFailureReason());
+      Assert.assertTrue(job.isWriterFactoryCreated(), "KIF repush should continue to the writer path");
+      Assert.assertEquals(quotaRefreshCount.get(), 0L, "KIF repush must not refresh or enforce storage quota");
+    }
+  }
+
+  private static long getTotalInputDataSize(QuotaTestingDataWriterSparkJob job) {
+    return job.getTaskTracker().getTotalKeySize() + job.getTaskTracker().getTotalValueSize();
+  }
+
+  private PushJobSetting getDefaultKafkaInputPushJobSetting() {
+    PushJobSetting setting = new PushJobSetting();
+    setting.storeName = Utils.getUniqueString("TEST_STORE");
+    setting.jobId = Utils.getUniqueString("TEST_JOB");
+    setting.isSourceKafka = true;
+    setting.kafkaInputTopic = Version.composeKafkaTopic(setting.storeName, 1);
+    setting.repushSourcePubsubBroker = "localhost:9092";
+    setting.repushTTLEnabled = false;
+    setting.topic = Version.composeKafkaTopic(setting.storeName, 2);
+    setting.pushDestinationPubsubBroker = "localhost:9092";
+    setting.partitionerClass = DefaultVenicePartitioner.class.getCanonicalName();
+    setting.partitionerParams = null;
+    setting.partitionCount = 1;
+    setting.storeKeySchema = Schema.create(Schema.Type.STRING);
+    setting.storeCompressionStrategy = CompressionStrategy.NO_OP;
+    setting.topicCompressionStrategy = CompressionStrategy.NO_OP;
+    setting.sourceKafkaInputVersionInfo = new VersionImpl(setting.storeName, 1, setting.jobId);
+    setting.sourceKafkaInputVersionInfo.setCompressionStrategy(CompressionStrategy.NO_OP);
+    setting.sourceVersionCompressionStrategy = CompressionStrategy.NO_OP;
+    setting.valueSchemaId = 1;
+    setting.inputHasRecords = true;
+    setting.inputFileDataSizeInBytes = 1000;
+    setting.maxRecordSizeBytes = Integer.MAX_VALUE;
+    return setting;
+  }
+
   private PushJobSetting getDefaultPushJobSetting(File inputDir, Schema dataSchema) {
     PushJobSetting setting = new PushJobSetting();
     setting.storeName = Utils.getUniqueString("TEST_STORE");
@@ -530,5 +654,67 @@ public class AbstractDataWriterSparkJobTest {
     protected Dataset<Row> getKafkaInputDataFrame() {
       return getSparkSession().emptyDataFrame();
     }
+  }
+
+  private static class QuotaTestingDataWriterSparkJob extends DataWriterSparkJob {
+    private boolean writerFactoryCreated = false;
+
+    boolean isWriterFactoryCreated() {
+      return writerFactoryCreated;
+    }
+
+    @Override
+    protected MapPartitionsFunction<Row, Row> createPartitionWriterFactory(
+        Broadcast<Properties> broadcastProperties,
+        com.linkedin.venice.spark.datawriter.task.DataWriterAccumulators accumulators) {
+      writerFactoryCreated = true;
+      return iterator -> {
+        long recordCount = 0;
+        int partitionId = 0;
+        while (iterator.hasNext()) {
+          Row row = iterator.next();
+          partitionId = row.getInt(row.fieldIndex(PARTITION_COLUMN_NAME));
+          recordCount++;
+          accumulators.outputRecordCounter.add(1);
+        }
+        accumulators.partitionWriterCloseCounter.add(1);
+        return Collections.singletonList(RowFactory.create(partitionId, recordCount)).iterator();
+      };
+    }
+  }
+
+  private static class KifQuotaTestingDataWriterSparkJob extends QuotaTestingDataWriterSparkJob {
+    @Override
+    protected Dataset<Row> getKafkaInputDataFrame() {
+      List<Row> rows = Collections.singletonList(
+          new GenericRowWithSchema(
+              new Object[] { "region1", 0, 100L, MessageType.PUT.getValue(), 1, "test-key".getBytes(),
+                  "test-value".getBytes(), 1, "rmd".getBytes(), null },
+              RAW_PUBSUB_INPUT_TABLE_SCHEMA));
+      return getSparkSession().createDataFrame(rows, RAW_PUBSUB_INPUT_TABLE_SCHEMA);
+    }
+  }
+
+  public enum StorageQuotaRefresh {
+    SMALL {
+      @Override
+      long resolve(QuotaTestingDataWriterSparkJob job) {
+        return 1L;
+      }
+    },
+    BIG {
+      @Override
+      long resolve(QuotaTestingDataWriterSparkJob job) {
+        return Long.MAX_VALUE;
+      }
+    },
+    EXACT_INPUT_SIZE {
+      @Override
+      long resolve(QuotaTestingDataWriterSparkJob job) {
+        return getTotalInputDataSize(job);
+      }
+    };
+
+    abstract long resolve(QuotaTestingDataWriterSparkJob job);
   }
 }
